refactor implement HttpResponseFactory
