+++ src/main/java/service/UserCreateController.java
@@ -4,6 +4,7 @@ import db.DataBase;
 import model.User;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
+import webserver.response.HttpResponseFactory;
 
 import java.util.Map;
 
@@ -13,6 +14,6 @@ public class UserCreateController extends GetController {
         Map<String, String> queryMap = httpRequest.getQueryMap();
         User user = User.from(queryMap);
         DataBase.addUser(user);
-        return HttpResponse.response302("/index.html");
+        return HttpResponseFactory.response302("/index.html");
     }
 }
