+++ src/main/java/webserver/response/HttpResponse.java
@@ -3,9 +3,6 @@ package webserver.response;
 import webserver.request.Headers;
 import webserver.request.Protocol;
 
-import java.util.HashMap;
-import java.util.Map;
-
 /**
  * 현재 DTO느낌으로 사용하고 있으나 오브젝트로 역할을 가져가야한다.
  */
@@ -28,20 +25,6 @@ public class HttpResponse {
         this.code = code;
         this.headers = Headers.empty();
         this.protocol = Protocol.defaultProtocol();
-    }
-
-    public static HttpResponse response202() {
-        return new HttpResponse(new byte[]{}, "202", Headers.empty(), Protocol.defaultProtocol());
-    }
-
-    public static HttpResponse response302(String location) {
-        Map<String, String> headerMap = new HashMap<>();
-        headerMap.put(Headers.LOCATION, location);
-        return new HttpResponse(new byte[]{}, "302", new Headers(headerMap), Protocol.defaultProtocol());
-    }
-
-    public static HttpResponse response404() {
-        return new HttpResponse(new byte[]{}, "404", Headers.empty(), Protocol.defaultProtocol());
     }
